+++ CHANGES.rst
@@ -10,6 +10,7 @@ Unreleased Changes
   * Retry MySQL DB creation during Docker tests until it succeeds, or fails 10 times.
   * Add testing of PhantomJS in Docker image testing; check version and that it actually works (GET a page).
   * More reliable stopping and removing of Docker containers during Docker image tests.
+* `Issue #63 <https://github.com/jantman/biweeklybudget/issues/63>`_ - Enable gunicorn request logging in Docker container.
 
 0.1.0 (2017-05-07)
 ------------------